feat: Логика работы с данными и datasource вынесена в AnalysisPresenter
